refactor: LottoShop 객체를 domain 패키지로 이관한다
ref: https://github.com/next-step/kotlin-lotto/pull/993#discussion_r1406283659
