+++ src/main/kotlin/lotto2/domain/LottoMoney.kt
@@ -1,7 +1,5 @@
 package lotto2.domain
 
-import lotto2.application.LottoShop
-
 data class LottoMoney(private var balance: Int) {
     init {
         require(balance >= MIN_LOTTO_MONEY) { "초기 금액은 0원 이상이어야 합니다." }
